AMP-Dimer Software/requirements.txt
torch
numpy
pandas
biopython
